Move material/nuclide related C API bindings into separate files
